Project Setup
Project was setup for Gradle with JUnit 5/Jupiter, Mockito (with
Hamcrest) and JaCoCo. The project will be split into three subprojects
and connected using TCP:

* The game, as the server.
* The UI, as a client.
* The computer opponent, as another client.

This structure should allow for swapping of clients and introduction of
new clients either from myself or third parties. So far the project is
only setup for the game server.