[WIP] Game Logic
Considered a change to how player actions are taken. Separating player
actions from the player class so that there is a much more intuitive way
of controlling what players can do at different states in the game. This
code has dustballed a bit and is also not tested. Commit made to backup
work for now and will also get the interactive rebase treatment.

The whole codebase feels quite bloated and with the looming challenge of
figuring out how to best handle Java threading as well as handling
network connections, I feel the the project will become immensely
difficult to work in the future. I may need to consider restarting the
project and approaching the project from a completely different angle.
This project will most likely go on hiatus for a while.